Moved future_returning and future_raising to a separate test helper file.
