+++ tests/handlers/test_healthcheck_handler.py
@@ -4,6 +4,7 @@ import tornado.concurrent
 import tornado.testing
 
 from run_service import make_app
+from tests.helpers import future_returning, future_raising
 
 
 class TestHealthcheckHandler(tornado.testing.AsyncHTTPTestCase):
@@ -18,37 +19,27 @@ class TestHealthcheckHandler(tornado.testing.AsyncHTTPTestCase):
             "mysql": self.mysql
         })
 
-    def future_returning(self, result):
-        future = tornado.concurrent.Future()
-        future.set_result(result)
-        return future
-
-    def future_raising(self, exception):
-        future = tornado.concurrent.Future()
-        future.set_exception(exception)
-        return future
-
     def mysql_execute_returning(self, result):
         mysql_cursor = mock.Mock()
-        mysql_cursor.execute.return_value = self.future_returning(result)
-        mysql_cursor.close.return_value = self.future_returning(None)
+        mysql_cursor.execute.return_value = future_returning(result)
+        mysql_cursor.close.return_value = future_returning(None)
 
-        self.mysql.cursor.return_value = self.future_returning(mysql_cursor)
+        self.mysql.cursor.return_value = future_returning(mysql_cursor)
 
         return mysql_cursor
 
     def mysql_execute_raising(self, exception):
         mysql_cursor = mock.Mock()
-        mysql_cursor.execute.return_value = self.future_raising(exception)
-        mysql_cursor.close.return_value = self.future_returning(None)
+        mysql_cursor.execute.return_value = future_raising(exception)
+        mysql_cursor.close.return_value = future_returning(None)
 
-        self.mysql.cursor.return_value = self.future_returning(mysql_cursor)
+        self.mysql.cursor.return_value = future_returning(mysql_cursor)
 
         return mysql_cursor
 
     def test_get_returns_200_when_healthy(self):
-        self.mongo_db.command.return_value = self.future_returning({"ok": 1.0})
-        self.redis.ping.return_value = self.future_returning(b"PONG")
+        self.mongo_db.command.return_value = future_returning({"ok": 1.0})
+        self.redis.ping.return_value = future_returning(b"PONG")
         mysql_cursor = self.mysql_execute_returning(None)
 
         response = self.fetch("/api/v1/healthcheck")
@@ -65,8 +56,8 @@ class TestHealthcheckHandler(tornado.testing.AsyncHTTPTestCase):
         mysql_cursor.close.assert_called_once_with()
 
     def test_get_returns_500_if_mongo_database_is_down(self):
-        self.mongo_db.command.return_value = self.future_raising(Exception("mongo error"))
-        self.redis.ping.return_value = self.future_returning(b"PONG")
+        self.mongo_db.command.return_value = future_raising(Exception("mongo error"))
+        self.redis.ping.return_value = future_returning(b"PONG")
         self.mysql_execute_returning(None)
 
         response = self.fetch("/api/v1/healthcheck")
@@ -74,8 +65,8 @@ class TestHealthcheckHandler(tornado.testing.AsyncHTTPTestCase):
         self.assertEqual(500, response.code)
 
     def test_get_returns_500_if_redis_is_down(self):
-        self.mongo_db.command.return_value = self.future_returning({"ok": 1.0})
-        self.redis.ping.return_value = self.future_raising(Exception("redis error"))
+        self.mongo_db.command.return_value = future_returning({"ok": 1.0})
+        self.redis.ping.return_value = future_raising(Exception("redis error"))
         self.mysql_execute_returning(None)
 
         response = self.fetch("/api/v1/healthcheck")
@@ -83,8 +74,8 @@ class TestHealthcheckHandler(tornado.testing.AsyncHTTPTestCase):
         self.assertEqual(500, response.code)
 
     def test_get_returns_500_if_mysql_is_down(self):
-        self.mongo_db.command.return_value = self.future_returning({"ok": 1.0})
-        self.redis.ping.return_value = self.future_returning(b"PONG")
+        self.mongo_db.command.return_value = future_returning({"ok": 1.0})
+        self.redis.ping.return_value = future_returning(b"PONG")
         mysql_cursor = self.mysql_execute_raising(Exception("mysql error"))
 
         response = self.fetch("/api/v1/healthcheck")
